+++ pkg/validation/validator.go
@@ -11,12 +11,16 @@ type Config struct {
 	G8sClient versioned.Interface
 	K8sClient kubernetes.Interface
 	Logger    micrologger.Logger
+
+	Provider string
 }
 
 type Validator struct {
 	g8sClient versioned.Interface
 	k8sClient kubernetes.Interface
 	logger    micrologger.Logger
+
+	provider string
 }
 
 func NewValidator(config Config) (*Validator, error) {
@@ -30,10 +34,16 @@ func NewValidator(config Config) (*Validator, error) {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
+	if config.Provider == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Provider must not be empty", config)
+	}
+
 	validator := &Validator{
 		g8sClient: config.G8sClient,
 		k8sClient: config.K8sClient,
 		logger:    config.Logger,
+
+		provider: config.Provider,
 	}
 
 	return validator, nil