+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `compatibleProvider` metadata validation.
+
 ## [4.5.0] - 2021-02-25
 
 ### Added